Move helper functions into separate modules and handle server and client env
